requirements.txt
asgiref==3.7.2
boto3==1.20.26
botocore==1.23.54
Django==4.2.2
dj-database-url==0.5.0
django-auto-logout==0.5.0
django-ckeditor==6.0.0
django-ckeditor-5==0.2.10
django-crispy-forms==2.0
django-dotenv==1.4.2
django-environ==0.9.0
django-recaptcha==3.0.0
django-filter==21.1
django-formset-js-improved==0.5.0.2
django-jazzmin==2.6.0
django-jquery-js==3.1.1
django-js-asset==1.2.2
django-mailgun-provider==0.2.3
django-plaintext-password==0.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-social-share==2.2.1
django-static-fontawesome==5.14.0.0
django-storages==1.12.3
django-taggit==3.0.0
django-import-export
django-templated-mail==1.1.1
django-tinymce==3.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.0.5
# Pillow==9.1.0
Pillow==10.1.0
python-decouple==3.5
rjsmin==1.1.0
s3transfer==0.5.2
validate-email==1.3
whitenoise==5.2.0	
shortuuid==1.0.11
django-mathfilters==1.0.0
django-admin-sortable2==2.1.10
stripe==7.0.0
# psycopg2-binary==2.9.9
paypal-payouts-sdk==1.0.0
paypalhttp==1.0.1
django-paypal==2.0
django-mailgun-provider==0.2.3
django-anymail==9.1
geoip2==4.6.0
requests==2.31.0
s3transfer==0.5.2
django-user-agents==0.4.0
gunicorn==21.2.0
channels==4.0.0
daphne==4.0.0
# django-heroku==0.3.1
crispy-bootstrap5==0.7
pandas==2.1.1
redis==5.0.1
django-theme-material-kit==1.0.19